// src/oasisQuery/oasis-nexus/api.ts
  export interface RuntimeEvent {
    network: Network
    layer: Runtime
  }